Improve domain modeling. Better weight input

+++ app/src/main/java/dev/sanastasov/bybon/weight/data/WeightEntryEntity.kt
@@ -8,5 +8,5 @@ import java.time.LocalDate
 @Entity("weight_entry")
 data class WeightEntryEntity(
     @PrimaryKey val date: LocalDate,
-    @ColumnInfo("weight") val weight: Float,
+    @ColumnInfo("weight") val weight: Int,
 )
